refactor(billing): extract usage panels
